refactor: move MetadataType to SMSConsts
