--- src/main/java/org/hisp/dhis/smscompression/models/MetadataType.java
@@ -1,16 +1,0 @@
-package org.hisp.dhis.smscompression.models;
-
-public enum MetadataType
-{
-    USER,
-    TRACKED_ENTITY_TYPE,
-    TRACKED_ENTITY_ATTRIBUTE,
-    PROGRAM,
-    ORGANISATION_UNIT,
-    DATA_ELEMENT,
-    CATEGORY_OPTION_COMBO,
-    DATASET,
-    PROGRAM_STAGE,
-
-    ;
-}
